requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.2
django-localflavor==3.1
django-phonenumber-field==7.0.2
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.0
phonenumberslite==8.13.8
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.2.2
python-stdnum==1.18
sqlparse==0.4.3
tomli==2.0.1
